Fix rotation of placeholders: Placeholders of a package are rotated about the origin of the package in counterclockwise direction if package has been flipped.
This rule also applies for the whole package: If flipped, the rotation is
counterclockwise.

It should be verified whether this rule is reasonable.
This rule seems in contradiction to the principle that the user sees "through"
the board all the time. Rotation of packages would be independent of
the flip status.

If this rule is not reasonable then the rotation operations for the package
and the placeholders must be changed.